% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dph_shiny_app.R
\name{new_dph_shiny_app}
\alias{new_dph_shiny_app}
\title{Function creates Shiny app project directory structured as an R package.}
\usage{
new_dph_shiny_app(path, ...)
}
\arguments{
\item{path}{The path where to create the project is supplied by the RStudio
interface.}

\item{...}{Passes other parameters from the RStudio new project wizard}
}
\description{
Function creates Shiny app project directory structured as an R package.
}
