% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_app_template.R
\name{use_app_template}
\alias{use_app_template}
\title{Scaffold shiny app in current working directory}
\usage{
use_app_template(template = "bslib_top_modules")
}
\arguments{
\item{template}{One of vanilla_shiny, shiny_dashboard_modules,
bslib_top_modules, module_template}
}
\description{
This function will create files and folders in the
user's current working directory...
}
